Se actualizaron templates y se implementó SSO. Se comenzó con el manejo de las aplicaciones clientes consumiendo el servicio de gestión de aplicaciones

+++ src/main/java/ar/gob/ambiente/servicios/gestiontramites/facades/EstadoFacade.java
@@ -66,7 +66,7 @@ public class EstadoFacade extends AbstractFacade<Estado> {
      * @param app: ID de la entidad
      * @return: True o False
      */
-    public List<Estado> getEstadosXapp(int app){
+    public List<Estado> getEstadosXapp(long app){
         em = getEntityManager();        
         String queryString = "SELECT est FROM Estado est "
                 + "WHERE est.app = :app "
